hoist import, move test to libtiff
